+++ src/test.task.ts
@@ -6,7 +6,7 @@ import type {Task} from './task/task.js';
 import {TaskError} from './task/task.js';
 import {toBuildOutPath, toRootPath} from './paths.js';
 import {SYSTEM_BUILD_NAME} from './build/defaultBuildConfig.js';
-import {loadGroConfig} from './config/config.js';
+import {loadConfig} from './config/config.js';
 import {buildSourceDirectory} from './build/buildSourceDirectory.js';
 
 // Runs the project's tests: `gro test [...args]`
@@ -28,7 +28,7 @@ export const task: Task = {
 		// TODO cleaner way to detect & rebuild?
 		if (!(await fs.exists(testsBuildDir))) {
 			const timingToLoadConfig = timings.start('load config');
-			const config = await loadGroConfig(fs, dev);
+			const config = await loadConfig(fs, dev);
 			timingToLoadConfig();
 
 			const timingToPrebuild = timings.start('prebuild');