+++ changelog.md
@@ -6,6 +6,8 @@
   ([#207](https://github.com/feltcoop/gro/pull/207))
 - **break**: build configs now fail to validate if any input path strings do not exist
   ([#207](https://github.com/feltcoop/gro/pull/207))
+- **break**: rename `loadConfig` from `loadGroConfig`
+  ([#207](https://github.com/feltcoop/gro/pull/207))
 - fix clean
   ([#207](https://github.com/feltcoop/gro/pull/207))